DailyShare/requirements.txt
Django~=6.0
djangorestframework~=3.16.1
gunicorn
pillow~=12.0.0
